Smooth out development environment setup (#17)
Dependencies and setup tasks are listed explicitly in README.md, and
installing requirements is simplified to one command. I also made the
whole package runnable as

$ spot

but that's mostly for convenience - when we're building a more
fleshed-out CLI, we'll probably want to clean up the package structure a
bit and change the entry point from main to a specific CLI script.