Use boto3 instead of custom code (#27)
Use boto3's Lambda client to invoke functions instead of a
manually-constructed POST.

The boto3 call is wrapped in a Future to keep the code consistent with
the existing structure and avoid potential side effects in Lambda's
asynchronous invocations. If the asynchronous invocations don't have
side effects, they should be used instead.

The POST was constructed using IAMAuth, which in turn depended on
AWSCredentialsFetch. Nothing else depended on these classes, so remove
them.

The boto3 call uses the function's name or ARN from the "application"
field in the workload instead of a URL constructed with the "host",
"stage", and "resource" fields. Update the documentation to reflect
this.