Fix commit failure in format action (#39)
By default, actions/checkout@v2 checks out the result of merging with the base branch. This prevents the workflow from committing and pushing changes. Check out the head commit instead.

git update-index was also failing, though it's not clear if that was related to the checkout issue. Avoid the problem by using black . --check.